Made everything more object oriented designed:
There are now controls and children of controls for both Human contorl and AI control.

Players now have children, each child will be a different playable character with their unique attacks and / or timings.

Each Player / Character has a controller class inside of it that handles the control (like above)